# pkg/broadcastwriter: avoid alloc w/ WriteString