@unclejack 2015-03-29
@jfrazelle I've added this to the 1.6.0 milestone because it's a very small change and it's useful. Please feel free to remove it from the milestone or tell me to do it if you think we shouldn't include this change.